Move files.rs and jobs.rs to transfer module
