+++ peernode/src/peer-node/client.rs
@@ -3,6 +3,7 @@ pub mod consumer;
 pub mod peer;
 pub mod producer;
 pub mod store;
+pub mod transfer;
 
 use std::io::{self, Write};
 use store::Configurations;
